combine the tests and only run on node 14+
