--- dev-packages/node-integration-tests/suites/tracing-experimental/mysql/withConnect/test.ts
@@ -1,31 +1,0 @@
-import { createRunner } from '../../../../utils/runner';
-
-const EXPECTED_TRANSACTION = {
-  transaction: 'Test Transaction',
-  spans: expect.arrayContaining([
-    expect.objectContaining({
-      description: 'SELECT 1 + 1 AS solution',
-      op: 'db',
-      data: expect.objectContaining({
-        'db.system': 'mysql',
-        'net.peer.name': 'localhost',
-        'net.peer.port': 3306,
-        'db.user': 'root',
-      }),
-    }),
-    expect.objectContaining({
-      description: 'SELECT NOW()',
-      op: 'db',
-      data: expect.objectContaining({
-        'db.system': 'mysql',
-        'net.peer.name': 'localhost',
-        'net.peer.port': 3306,
-        'db.user': 'root',
-      }),
-    }),
-  ]),
-};
-
-test('should auto-instrument `mysql` package when using connection.connect()', done => {
-  createRunner(__dirname, 'scenario.js').expect({ transaction: EXPECTED_TRANSACTION }).start(done);
-});
